+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Added a collection of theme colors to be available in the editor.
+- Added alternative style for Media & Text block.
 
 ## 0.1.0